🚚 resource(dynamics): move private files
