--- src/galax/dynamics/_src/dynamics/__init__.py
@@ -1,14 +1,0 @@
-"""General Dynamics. Private module."""
-
-__all__ = [
-    "OrbitSolver",
-    # Fields
-    "AbstractOrbitField",
-    "HamiltonianField",
-    "NBodyField",
-]
-
-from .field_base import AbstractOrbitField
-from .field_hamiltonian import HamiltonianField
-from .field_nbody import NBodyField
-from .solver import OrbitSolver
