+++ src/galax/dynamics/_src/legacy/funcs.py
@@ -17,8 +17,7 @@ import galax.coordinates as gc
 import galax.dynamics._src.custom_types as gdt
 import galax.potential as gp
 from .integrator import Integrator
-from galax.dynamics._src.dynamics import HamiltonianField
-from galax.dynamics._src.orbit import Orbit
+from galax.dynamics._src.orbit import HamiltonianField, Orbit
 
 # TODO: enable setting the default integrator
 default_integrator: Integrator = Integrator()
