+++ src/galax/dynamics/_src/orbit/__init__.py
@@ -3,17 +3,28 @@
 __all__ = [
     "AbstractOrbit",
     "Orbit",
+    # Solve orbits
     "compute_orbit",
+    "OrbitSolver",
+    # Fields
+    "AbstractOrbitField",
+    "HamiltonianField",
+    "NBodyField",
+    # misc
     "plot_components",
     "PhaseSpaceInterpolation",
 ]
 
 from .base import AbstractOrbit
 from .compute import compute_orbit
+from .field_base import AbstractOrbitField
+from .field_hamiltonian import HamiltonianField
+from .field_nbody import NBodyField
 from .interp import PhaseSpaceInterpolation
 from .orbit import Orbit
 from .plot_helper import plot_components
+from .solver import OrbitSolver
 
 # Register by import
 # isort: split
-from . import register_dfx  # noqa: F401
+from . import register_dfx, register_gc  # noqa: F401
