+++ src/galax/dynamics/fields.py
@@ -12,12 +12,10 @@ from jaxtyping import install_import_hook
 from galax.setup_package import RUNTIME_TYPECHECKER
 
 with install_import_hook("galax.dynamics.fields", RUNTIME_TYPECHECKER):
-    from ._src.dynamics import (
-        AbstractOrbitField,
-        HamiltonianField,
-        NBodyField,
-    )
     from ._src.fields import AbstractField
+    from ._src.orbit.field_base import AbstractOrbitField
+    from ._src.orbit.field_hamiltonian import HamiltonianField
+    from ._src.orbit.field_nbody import NBodyField
 
 # Cleanup
 del install_import_hook, RUNTIME_TYPECHECKER
